poemgr: configure power_budget to override detected limit
If the new power_budget setting is set to anything larger than zero,
we will ignore the official power budget based on the supply 802.3 type.

This is useful when a bunch of powered devices slightly peak when
they all boot at the same time and we know that total consumption
will otherwise be fine.

Tested with an 802.3at injector as supply and four active-to-passive-PoE
converters in front of three sxtsq-5-ac and one litebeam-ac-gen2.
The official power budget of 21W for 802.3at supply only consistently
powered on one device, often two, and rarely three.

With power_budget=24 all four devices consistently boot up.