pd69104: Avoid self-induced pointer casts
As a general rule, pointer casts in C are dangerous. While there are
valid uses, there are many more uses where the behavior is undefined.
A good strategy is to avoid pointer casts where possible.

What is the "data" argument to i2c_smbus_access()? Is it an array of
bytes, a string, or a memory region? This declaration is ambiguous.
Resolving this is trivial: Do not cast 'data' to a (char*), and just
leave it as a (union i2c_smbus_data*). Not only is this unambiguous,
it also maintains type safety throughout the call stack.