poemgr: Fix compiler warnings in poemgr.c
First, usleep() was undeclared. This is fixed by #including <unistd.h>

Second, in poemgr_load_port_settings(), move variable declarations to
the top of the function, in reverse christmas tree order. There is a
popular belief that minimally scoping variables provides some
performance advantage. There is no evidence of this with modern
compilers. I am going for readability, which is a more realistic
concer, given that "struct uci_section *s" was declared twice.

Third, remove "disabled" from poemgr_load_settings(). It is not used.

Fourth, poemgr_load_settings() has a bug. If any of the first three
gotos are hit, the call to json_object_put(root_obj) will have an
uninitialized pointer, "root_obj", likely resulting in a crash.
To resolve this, replace the gotos with an elegant "return ret".