Makefile: compile object files before linking
Instead of compiling poemgr directly in one step, have intermediate object
files. This makes can make it a little bit easier to handle when more files
are added in the future.

But this also requires that the compiler outputs dependencies between the
files to ensure that make can figure out when a files needs to be
recompiled.

It also has the nice side effect that poemgr itself will be rebuild
whenever the system headers (it includes) are changed.

Signed-off-by: Sven Eckelmann <sven@narfation.org>